split multi role
